requirements.txt
pymysql
pyDatalog
cryptography
python-dotenv